Moved libz loading to module level, rather than inside class constructor.
I noticed, using this with Spigot, that execve("/sbin/ldconfig", "-p") was
being called very frequently. Turns out that ctypes.utils shells out to that
command to find a shared library to load. Loading it in the constructor
meant that this program was being run every time a GzipChunk object was
created.

We only actually need to load the shared library once, at module inclusion
time.